Split CMakeList.txt to use individual sub-directory build for lib, exe, test
