+++ src/file_filter.cc
@@ -1,5 +1,5 @@
 
-#include "libsrc/filter.h"
+#include <filter.h>
 
 #include <filesystem>
 #include <fstream>
