+++ tmpl_version.h.in
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "libsrc/proginfo.h"
+#include <proginfo.h>
 
 [[maybe_unused]] static constexpr std::string_view PROGRAM_INFO_GENERATED =
   "Program information - Generated by cmake @CMAKE_VERSION@ on @BUILD_TIMESTAMP@";
